Split refresh coordinator modules
